Do not intern descriptors and constants
These objects are usually either short lived (e.g. the lifespan of method creation) or very long lived (e.g. stored in a constant field). Limiting them to the lifespan of the Gizmo instance does not make a lot of sense in this perspective. But, leave the door open to allow us to later explicitly intern these constants.

+++ src/main/java/io/quarkus/gizmo2/impl/constant/IntConstant.java
@@ -17,7 +17,7 @@ public final class IntConstant extends IntBasedConstant {
     }
 
     public int intValue() {
-        return value;
+        return value.intValue();
     }
 
     public boolean equals(final ConstantImpl obj) {
